travail sur les quotes afin que quand il y ait des quotes cela fasse bien 1 seul token. Aussi mise au propre du parser avec un decoupage en fonctions du tokenizer et ajout d'une categorie quote
